+++ pkg/services/session/storage/types.go
@@ -11,6 +11,16 @@ type PrivateToken struct {
 	exp uint64
 }
 
+// SetSessionKey sets a private session key.
+func (t *PrivateToken) SetSessionKey(sessionKey *ecdsa.PrivateKey) {
+	t.sessionKey = sessionKey
+}
+
+// SetExpiredAt sets epoch number until token is valid.
+func (t *PrivateToken) SetExpiredAt(exp uint64) {
+	t.exp = exp
+}
+
 // SessionKey returns the private session key.
 func (t *PrivateToken) SessionKey() *ecdsa.PrivateKey {
 	return t.sessionKey
